# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  version: 1.0.0
  title: Product API
  description: Apis to manage vendors and customers

paths:
  /api/vendor/signup:
    post:
      tags:
        - Vendor
      summary: Vendor signup
      operationId: vendorSignup
      description: Vendor signup
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorSignup"
      responses:
        '200':
          description: Successful response
        '400':
          description: Failed response
  /api/vendor/login:
    post:
      tags:
        - Vendor
      summary: Vendor login
      operationId: vendorLogin
      description: Vendor login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorLogin"
      responses:
        '200':
          description: Successful response
        '400':
          description: Failed response
  /api/vendor/me:
    get:
      tags:
        - Vendor
      summary: Get vendor info
      operationId: vendorMe
      description: Get vendor info
      responses:
        '200':
          description: Successful response

components:
  schemas:
    VendorSignup:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          description: Email address of vendor
          example: example@email.com
        password:
          type: string
          description: Password of vendor
          example: password1234
        name:
          type: string
          description: name of vendor
          example: joybox games
    VendorLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          description: Email address of vendor
          example: example@email.com
        password:
          type: string
          description: Password of vendor
          example: password1234